exposome_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=2.1.0=pypi_0
anyio=4.2.0=py311hca03da5_0
appnope=0.1.2=py311hca03da5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py311h80987f9_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
async-lru=2.0.4=py311hca03da5_0
attrs=23.1.0=py311hca03da5_0
babel=2.11.0=py311hca03da5_0
beautifulsoup4=4.12.2=py311hca03da5_0
biopython=1.83=pypi_0
bleach=4.1.0=pyhd3eb1b0_0
brotli-python=1.0.9=py311h313beb8_7
bzip2=1.0.8=h80987f9_5
ca-certificates=2023.12.12=hca03da5_0
cartographs=2.0.0=pypi_0
certifi=2024.2.2=py311hca03da5_0
cffi=1.16.0=py311h80987f9_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colormath=3.0.0=pypi_0
comm=0.2.1=py311hca03da5_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py311h313beb8_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dm-tree=0.1.8=pypi_0
et-xmlfile=1.1.0=pypi_0
executing=0.8.3=pyhd3eb1b0_0
flatbuffers=24.3.7=pypi_0
fonttools=4.49.0=pypi_0
fuzzywuzzy=0.18.0=pypi_0
gast=0.5.4=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.62.1=pypi_0
gseapy=1.1.2=pypi_0
h5py=3.10.0=pypi_0
idna=3.4=py311hca03da5_0
infomap=2.7.1=pypi_0
ipykernel=6.28.0=py311hca03da5_0
ipython=8.20.0=py311hca03da5_0
jedi=0.18.1=py311hca03da5_1
jinja2=3.1.3=py311hca03da5_0
joblib=1.3.2=pypi_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.19.2=py311hca03da5_0
jsonschema-specifications=2023.7.1=py311hca03da5_0
jupyter-lsp=2.2.0=py311hca03da5_0
jupyter_client=8.6.0=py311hca03da5_0
jupyter_core=5.5.0=py311hca03da5_0
jupyter_events=0.8.0=py311hca03da5_0
jupyter_server=2.10.0=py311hca03da5_0
jupyter_server_terminals=0.4.4=py311hca03da5_1
jupyterlab=4.0.11=py311hca03da5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.25.1=py311hca03da5_0
kaleido=0.2.1=pypi_0
keras=3.0.5=pypi_0
kiwisolver=1.4.5=pypi_0
libclang=16.0.6=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_0
libsodium=1.0.18=h1a28f6b_0
llvmlite=0.42.0=pypi_0
markdown=3.5.2=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py311h80987f9_0
matplotlib=3.8.3=pypi_0
matplotlib-inline=0.1.6=py311hca03da5_0
mdurl=0.1.2=pypi_0
mistune=2.0.4=py311hca03da5_0
ml-dtypes=0.3.2=pypi_0
namex=0.0.7=pypi_0
nbclient=0.8.0=py311hca03da5_0
nbconvert=7.10.0=py311hca03da5_0
nbformat=5.9.2=py311hca03da5_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=py311hca03da5_0
networkx=3.2.1=pypi_0
notebook-shim=0.2.3=py311hca03da5_0
numba=0.59.0=pypi_0
numpy=1.26.4=pypi_0
openpyxl=3.1.2=pypi_0
openssl=3.0.13=h1a28f6b_0
opt-einsum=3.3.0=pypi_0
overrides=7.4.0=py311hca03da5_0
packaging=23.1=py311hca03da5_0
pandas=1.5.3=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
patsy=0.5.6=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pillow=10.2.0=pypi_0
pip=23.3.1=py311hca03da5_0
platformdirs=3.10.0=py311hca03da5_0
plotly=5.20.0=pypi_0
prometheus_client=0.14.1=py311hca03da5_0
prompt-toolkit=3.0.43=py311hca03da5_0
prompt_toolkit=3.0.43=hd3eb1b0_0
protobuf=4.25.3=pypi_0
psutil=5.9.0=py311h80987f9_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pubchempy=1.0.4=pypi_0
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py311hca03da5_1
pynndescent=0.5.11=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py311hca03da5_0
python=3.11.8=hb885b13_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py311hca03da5_0
python-json-logger=2.0.7=py311hca03da5_0
python-louvain=0.16=pypi_0
pytz=2023.3.post1=py311hca03da5_0
pyyaml=6.0.1=py311h80987f9_0
pyzmq=25.1.2=py311h313beb8_0
rdkit-pypi=2022.9.5=pypi_0
readline=8.2=h1a28f6b_0
referencing=0.30.2=py311hca03da5_0
requests=2.31.0=py311hca03da5_1
rfc3339-validator=0.1.4=py311hca03da5_0
rfc3986-validator=0.1.1=py311hca03da5_0
rich=13.7.1=pypi_0
rpds-py=0.10.6=py311hf0e4da2_0
scikit-learn=1.4.1.post1=pypi_0
scipy=1.12.0=pypi_0
seaborn=0.13.2=pypi_0
send2trash=1.8.2=py311hca03da5_0
setuptools=68.2.2=py311hca03da5_0
shapely=2.0.3=pypi_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=py311hca03da5_0
soupsieve=2.5=py311hca03da5_0
sqlite=3.41.2=h80987f9_0
stack_data=0.2.0=pyhd3eb1b0_0
statannotations=0.6.0=pypi_0
statsmodels=0.14.1=pypi_0
tenacity=8.2.3=pypi_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.16.1=pypi_0
tensorflow-io-gcs-filesystem=0.36.0=pypi_0
termcolor=2.4.0=pypi_0
terminado=0.17.1=py311hca03da5_0
threadpoolctl=3.3.0=pypi_0
tinycss2=1.2.1=py311hca03da5_0
tk=8.6.12=hb8d0fd4_0
tornado=6.3.3=py311h80987f9_0
tqdm=4.66.2=pypi_0
traitlets=5.7.1=py311hca03da5_0
typing-extensions=4.9.0=py311hca03da5_1
typing_extensions=4.9.0=py311hca03da5_1
tzdata=2024.1=pypi_0
umap-learn=0.5.5=pypi_0
urllib3=2.1.0=py311hca03da5_1
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py311hca03da5_1
websocket-client=0.58.0=py311hca03da5_4
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py311hca03da5_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h80987f9_0
yaml=0.2.5=h1a28f6b_0
zeromq=4.3.5=h313beb8_0
zlib=1.2.13=h5a0b063_0
